refactor(script-ductus): extract handwriting into a package the books can use
"Ductus" is the movement of the pen: not what a letter looks like, but the
order and direction of the strokes that make it. strokes.ts, truetype.ts
and ductusview.ts are the three files that know it -- the authored pen
paths, the real glyph outline read out of the shipped font, and the
filmstrip that joins them -- and they lived inside the app.

Nothing under code/packages/ may depend on something under code/programs/,
so the book generator could not import them at all. The other consumer
that wants filmstrips -- as printed figures rather than a live SVG --
simply could not reach them. ductusview.ts's own header already
anticipated this: "the book pipeline can take the serialised string
instead." Now it can.

WHAT MOVED, AND ONE THING THAT MOVED WITH IT

The three modules and all three test files, unchanged. The font
verification is the point of the package and none of it was loosened:
fractionOnInk requires an authored path to lie on the font's own ink,
consecutive segments must meet, and the strokes together must cover the
glyph rather than tracing a convenient part of it. A pen path that has
drifted from the letter it claims to draw fails a test rather than
shipping a plausible-looking lie.

data.ts moved too, as scriptdata.ts, along with the Letter and ScriptData
types. That was not obvious and it is right: a letter's stroke order is
verified against the very font its ScriptData names, so a test can assert
the two agree only if both live in one package. types.ts in the app
re-exports them, so every existing `from "./types.ts"` import is
unchanged -- and its old comment, which explained the types as a
deliberate duplicate, is no longer true and no longer there.

CHECK:BUNDLE EARNED ITS KEEP

vite.config.ts splits a "handwriting-tools" chunk so the pen paths and the
font parser do not sit in the interactive shell -- and it matched them BY
PATH. Moving the files silently emptied the chunk, which would have pulled
7,600 lines of handwriting code into the shell instead of loading it when
a learner opens a letter. Nothing else caught it: typecheck passed, all
725 app tests passed, the build succeeded. Only check:bundle failed, with
"expected one handwriting-tools chunk, found 0".

scriptdata is deliberately NOT in that chunk. The shell needs SCRIPTS on
first paint; the pen paths it does not.

JSDOM CAME ALONG, DELIBERATELY

Two tests hand the SVG serialiser's output to a REAL parser and assert
that a hostile caption cannot break out of an attribute or smuggle in a
<script>. A string comparison would pass on markup no browser accepts,
which is exactly the bug those two tests exist to catch. So the
environment moved with them rather than the tests being weakened to fit a
Node-only config.

VERIFIED

845 tests in the new package. The app: typecheck clean, 725 tests, build,
and check:bundle all pass. build-tool -validate-build-files reports no
missing prerequisite refs, with the new package chain-installed in
language-ladder's BUILD before its own npm install.

Closes HL-C114.
