+++ code/packages/typescript/script-ductus/BUILD
@@ -1,0 +1,2 @@
+npm install --silent
+npx vitest run --coverage
