+++ code/packages/typescript/script-ductus/src/ductusview.ts
