+++ code/packages/typescript/script-ductus/src/strokes.ts
