+++ code/packages/typescript/script-ductus/src/truetype.ts
