+++ code/programs/typescript/language-ladder/BUILD
@@ -30,6 +30,7 @@ set -e
 (cd ../../../packages/typescript/paint-vm && npm ci --silent)
 (cd ../../../packages/typescript/paint-vm-svg && npm ci --silent)
 (cd ../../../packages/typescript/human-language-data && npm ci --silent)
+(cd ../../../packages/typescript/script-ductus && npm install --silent)
 npm install
 npm run typecheck
 npm run build
